Move twin code into a separate class
