+++ verifier/src/main/java/com/aws/jverify/verifier/compiler/TopLevelDeclCompiler.java
@@ -1,0 +1,11 @@
+package com.aws.jverify.verifier.compiler;
+
+import com.aws.jverify.Nullable;
+import com.aws.jverify.generated.TopLevelDecl;
+
+import java.util.List;
+
+public interface TopLevelDeclCompiler {
+    @Nullable
+    List<TopLevelDecl> compile(TopLevelDecl clazz);
+}
