+++ verifier/src/main/java/com/aws/jverify/verifier/compiler/extend/dafny/CompileClassesExtendingClasses.java
@@ -1,0 +1,121 @@
+package com.aws.jverify.verifier.compiler.extend.dafny;
+
+import com.aws.jverify.Modifiable;
+import com.aws.jverify.Nullable;
+import com.aws.jverify.generated.*;
+import com.aws.jverify.verifier.compiler.ClassCompiler;
+import com.aws.jverify.verifier.compiler.JavaToDafnyCompiler;
+import com.aws.jverify.verifier.compiler.TopLevelDeclCompiler;
+import com.sun.tools.javac.code.Symbol;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class CompileClassesExtendingClasses implements TopLevelDeclCompiler {
+    ClassCompiler classCompiler;
+
+    public CompileClassesExtendingClasses(ClassCompiler classCompiler) {
+        this.classCompiler = classCompiler;
+    }
+
+    @Override
+    public @Nullable List<TopLevelDecl> compile(TopLevelDecl clazz) {
+        if (clazz instanceof ClassDecl classDecl) {
+            Symbol.ClassSymbol symbol = null;
+            return buildTraitAndClassTwin(symbol, classDecl.getOrigin(),
+                    classDecl.getNameNode(), classDecl.getMembers(), 
+                    classDecl.getTypeArgs(), classDecl.getTraits());
+        }
+        return List.of(clazz);
+    }
+    
+    /**
+     * Translating Java classes to both a Dafny trait and a class is used to support classes extending classes
+     */
+    private List<TopLevelDecl> buildTraitAndClassTwin(Symbol.ClassSymbol classSymbol,
+                                                       IOrigin origin, Name name,
+                                                       List<MemberDecl> members,
+                                                       List<TypeParameter> typeParameters,
+                                                       List<Type> superTraits) {
+        var traitMembers = new ArrayList<MemberDecl>();
+        var classMembers = new ArrayList<MemberDecl>();
+        var classNeeded = !JavaToDafnyCompiler.isInterfaceOrAbstract(classSymbol);
+
+        for(var member : members) {
+            switch (member) {
+                case Method method when !method.getHasStaticKeyword() -> {
+                    traitMembers.add(member);
+                    if (method.getBody() == null) {
+                        classMembers.add(member);
+                        // A bodyless trait in Dafny is abstract. 
+                        // You can not declare an assumed member in traits in Dafny
+                        // We add the assumed member to the class
+                    }
+                }
+
+                case Function function -> {
+                    traitMembers.add(function);
+                    if (function.getBody() == null) {
+                        // A bodyless trait in Dafny is abstract. 
+                        // You can not declare an assumed member in traits in Dafny
+                        // We add the assumed member to the class
+                        classMembers.add(member);
+                    }
+                }
+                case Constructor constructor -> {
+                    classNeeded = true;
+                    Method initMethod = constructorToInitMethod(name.getValue(), constructor);
+                    if (initMethod != null) {
+                        traitMembers.add(initMethod);
+                    }
+
+                    var classConstructor = new Constructor(constructor.getOrigin(), constructor.getNameNode(), null, false, null,
+                            constructor.getTypeArgs(), constructor.getIns(),
+                            constructor.getReq(), constructor.getEns(), constructor.getReads(),
+                            constructor.getDecreases(), constructor.getMod(),
+                            null);
+                    classMembers.add(classConstructor);
+                }
+                default -> traitMembers.add(member);
+            }
+        }
+
+        if (!JavaToDafnyCompiler.isInterface(classSymbol) || classCompiler.compiler.isAnnotated(classSymbol.type, Modifiable.class)) {
+            superTraits.add(new UserDefinedType(origin, new NameSegment(origin, "object", null)));
+        }
+
+        var trait = new TraitDecl(origin, name, null, typeParameters, traitMembers, superTraits, false);
+        List<Type> typeArgs = typeParameters.stream().map(
+                p -> (Type)new UserDefinedType(p.getOrigin(),
+                        new NameSegment(p.getOrigin(), p.getNameNode().getValue(), null))).toList();
+
+        if (classNeeded) {
+            var clazz = new ClassDecl(origin, new Name(name.getOrigin(), classCompiler.compiler.nameCompiler.CLASS_PREFIX + name.getValue()), null,
+                    typeParameters, classMembers, List.of(new UserDefinedType(origin, new NameSegment(origin, name.getValue(), typeArgs))), false);
+            return List.of(trait, clazz);
+        } else {
+            return List.of(trait);
+        }
+    }
+
+    /**
+     * To support 'super(...)' calls, we translate each Java constructor to an 'init' method in the Dafny trait
+     * The Dafny class constructor then calls the init method of the related trait, and of the trait of its parent type.
+     */
+    private Method constructorToInitMethod(String className, Constructor constructor) {
+        if (constructor.getBody() == null) {
+            return null;
+        }
+        BlockStmt body = new BlockStmt(constructor.getBody().getOrigin(), null, List.of(),
+                constructor.getBody().getBodyInit());
+        Name nameNode = new Name(constructor.getNameNode().getOrigin(), classCompiler.compiler.nameCompiler.getInitMethodName(className, constructor.getNameNode().getValue()));
+        var frameExpressions = new ArrayList<>(constructor.getMod().getExpressions());
+        var modClause = new Specification<>(frameExpressions, constructor.getMod().getAttributes());
+        frameExpressions.add(new FrameExpression(constructor.getOrigin(), new ThisExpr(constructor.getOrigin()), null));
+        return new Method(constructor.getOrigin(), nameNode, constructor.getAttributes(),
+                constructor.getIsGhost(), constructor.getSignatureEllipsis(), constructor.getTypeArgs(), constructor.getIns(),
+                constructor.getReq(), constructor.getEns(), constructor.getReads(), constructor.getDecreases(),
+                modClause,
+                false, List.of(), body, false);
+    }
+}
